feat: add status metric to detect failed runs (#10)
Previously we couldn't identify if a run failed just based on the
reported nodes. Because these were based on the contents of the files.
If a run would fail, the file could still exist and it would report a
valid ENR count.

This additional metric should make it easy to spot if a run failed.